Turned the downloader into a daemon like process that can just always be running in a loop

+++ current_data.py
@@ -1,5 +1,13 @@
 import urllib
 import datetime
+import os
+import time
 
-current_file_name = datetime.datetime.now().strftime('%Y%m%d_%H%M') + '.json'
-urllib.urlretrieve ("http://divvybikes.com/stations/json", current_file_name)
+while True:
+	current_file_name = datetime.datetime.now().strftime('%Y%m%d_%H%M') + '.json'
+	if not os.path.exists(current_file_name):
+		print "Downloading file... then sleeping."
+		urllib.urlretrieve ("http://divvybikes.com/stations/json", current_file_name)
+	else:
+		print "File already exists. Sleeping"
+	time.sleep(30)
